🛡️ Sentinel: [security improvement] Harden network security and prevent info leakage
- Migrated GeoIP lookup to HTTPS (freeipapi.com) and updated tests/flows.
- Enforced HTTPS by disabling cleartext traffic in production.
- Prevented information leakage (CWE-209) by hardening VpnError messages.
- Disabled application backups for better data protection.
- Optimized SettingsScreen UI for Maestro E2E reliability.
- Added VpnErrorTest.kt and documented security learnings.
