+++ app/src/debug/AndroidManifest.xml
@@ -15,7 +15,6 @@
         tools:ignore="ProtectedPermissions" />
 
     <application
-        android:name=".MainApplication"
         android:usesCleartextTraffic="true"
         android:networkSecurityConfig="@xml/network_security_config"
         android:allowBackup="true"
