+++ app/src/main/java/com/multiregionvpn/ui/components/VpnHeaderBar.kt
@@ -61,6 +61,7 @@ fun VpnHeaderBar(
                 // App name
                 Text(
                     text = "Region Router",
+                    modifier = Modifier.testTag("Region Router"),
                     style = MaterialTheme.typography.titleLarge,
                     fontWeight = FontWeight.SemiBold
                 )
@@ -134,4 +135,3 @@ enum class VpnStatus(val displayText: String) {
     DISCONNECTED("Disconnected"),
     ERROR("Error")
 }
-
